refactor(bulk): use params for create options notifier
