+++ lib/core/downloads/bulks.dart
@@ -1,5 +1,5 @@
 export 'bulks/notifications/bulk_download_notification.dart';
 export 'bulks/pages/bulk_download_page.dart';
 export 'bulks/widgets/bulk_download_task_tile.dart';
-export 'bulks/pages/create_bulk_download_task_sheet.dart';
+export 'bulks/pages/create_download_options_sheet.dart';
 export 'bulks/providers/providers.dart';
